more WIP for GPS binary position messaging
